Complete massive CommandHandler refactor following Discord.js best practices
MAJOR ACHIEVEMENT: Reduced CommandHandler.js from 1,045 to 299 lines (-746 lines, 71% reduction)

**All Commands Extracted:**
✅ /stop (106 lines) → src/commands/stop.js
✅ /transcribe (232 lines) → src/commands/transcribe.js
✅ /join (81 lines) → src/commands/join.js
✅ /summarize (171 lines) → src/commands/summarize.js
✅ /list (106 lines) → src/commands/list.js
✅ /last_recording (32 lines) → src/commands/last_recording.js
✅ /help → src/commands/help.js
✅ /ping → src/commands/ping.js

**Architecture Transformation:**
- CommandHandler.js now only contains core infrastructure:
  * Command registration and loading
  * Interaction routing and validation
  * Autocomplete handling
  * Background job management integration
  * Error handling middleware

- Individual commands are self-contained modules with:
  * Clear single responsibility
  * Dependency injection for services
  * Proper error handling
  * Timeout management
  * Helper function organization

**Benefits Achieved:**
- Follows Discord.js community best practices for command organization
- Dramatically improved maintainability and readability
- Each command can be tested and modified independently
- Easier onboarding for new developers
- Better separation of concerns
- Reduced cognitive load when working on specific features

**Technical Implementation:**
- Dynamic command loading with dependency injection
- Service wrapper pattern for external commands
- Preserved all advanced functionality (background jobs, timeouts, caching)
- Maintained backward compatibility
- Zero functionality lost during refactor

This transformation brings the Discord bot architecture in line with
industry standards while preserving its innovative background processing
capabilities and advanced voice recording features.
